Met en commun les modèles et vues de pièces entre le contrôle et le tri.
