+++ src/situations/tri/data/pieces.js
@@ -1,52 +1,57 @@
+const scene = {
+  largeur: 1008,
+  hauteur: 566
+};
+
 const pieces = [
-  { image: require('tri/assets/bonbon1.png'), posY: 433, posX: 196 },
-  { image: require('tri/assets/bonbon2.png'), posY: 433, posX: 723 },
-  { image: require('tri/assets/bonbon3.png'), posY: 447, posX: 91 },
-  { image: require('tri/assets/bonbon4.png'), posY: 468, posX: 404 },
-  { image: require('tri/assets/bonbon5.png'), posY: 452, posX: 457 },
-  { image: require('tri/assets/bonbon6.png'), posY: 430, posX: 301 },
-  { image: require('tri/assets/bonbon7.png'), posY: 435, posX: 450 },
-  { image: require('tri/assets/bonbon8.png'), posY: 434, posX: 688 },
-  { image: require('tri/assets/bonbon9.png'), posY: 471, posX: 605 },
-  { image: require('tri/assets/bonbon10.png'), posY: 450, posX: 638 },
-  { image: require('tri/assets/bonbon11.png'), posY: 474, posX: 402 },
-  { image: require('tri/assets/bonbon12.png'), posY: 448, posX: 501 },
-  { image: require('tri/assets/bonbon1.png'), posY: 477, posX: 316 },
-  { image: require('tri/assets/bonbon2.png'), posY: 443, posX: 706 },
-  { image: require('tri/assets/bonbon3.png'), posY: 458, posX: 308 },
-  { image: require('tri/assets/bonbon4.png'), posY: 477, posX: 664 },
-  { image: require('tri/assets/bonbon5.png'), posY: 473, posX: 654 },
-  { image: require('tri/assets/bonbon6.png'), posY: 445, posX: 387 },
-  { image: require('tri/assets/bonbon7.png'), posY: 471, posX: 440 },
-  { image: require('tri/assets/bonbon8.png'), posY: 476, posX: 713 },
-  { image: require('tri/assets/bonbon9.png'), posY: 445, posX: 231 },
-  { image: require('tri/assets/bonbon10.png'), posY: 455, posX: 140 },
-  { image: require('tri/assets/bonbon11.png'), posY: 466, posX: 750 },
-  { image: require('tri/assets/bonbon12.png'), posY: 456, posX: 671 },
-  { image: require('tri/assets/bonbon1.png'), posY: 477, posX: 795 },
-  { image: require('tri/assets/bonbon2.png'), posY: 464, posX: 453 },
-  { image: require('tri/assets/bonbon3.png'), posY: 430, posX: 855 },
-  { image: require('tri/assets/bonbon4.png'), posY: 463, posX: 230 },
-  { image: require('tri/assets/bonbon5.png'), posY: 469, posX: 328 },
-  { image: require('tri/assets/bonbon6.png'), posY: 475, posX: 374 },
-  { image: require('tri/assets/bonbon7.png'), posY: 437, posX: 474 },
-  { image: require('tri/assets/bonbon8.png'), posY: 476, posX: 106 },
-  { image: require('tri/assets/bonbon9.png'), posY: 453, posX: 619 },
-  { image: require('tri/assets/bonbon10.png'), posY: 462, posX: 303 },
-  { image: require('tri/assets/bonbon11.png'), posY: 443, posX: 243 },
-  { image: require('tri/assets/bonbon12.png'), posY: 448, posX: 174 },
-  { image: require('tri/assets/bonbon1.png'), posY: 471, posX: 632 },
-  { image: require('tri/assets/bonbon2.png'), posY: 439, posX: 289 },
-  { image: require('tri/assets/bonbon3.png'), posY: 456, posX: 346 },
-  { image: require('tri/assets/bonbon4.png'), posY: 474, posX: 808 },
-  { image: require('tri/assets/bonbon5.png'), posY: 450, posX: 281 },
-  { image: require('tri/assets/bonbon6.png'), posY: 433, posX: 335 },
-  { image: require('tri/assets/bonbon7.png'), posY: 470, posX: 751 },
-  { image: require('tri/assets/bonbon8.png'), posY: 443, posX: 403 },
-  { image: require('tri/assets/bonbon9.png'), posY: 451, posX: 832 },
-  { image: require('tri/assets/bonbon10.png'), posY: 450, posX: 255 },
-  { image: require('tri/assets/bonbon11.png'), posY: 451, posX: 299 },
-  { image: require('tri/assets/bonbon12.png'), posY: 450, posX: 51 }
+  { image: require('tri/assets/bonbon1.png'), y: 433 / scene.hauteur * 100, x: 196 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon2.png'), y: 433 / scene.hauteur * 100, x: 723 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon3.png'), y: 447 / scene.hauteur * 100, x: 91 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon4.png'), y: 468 / scene.hauteur * 100, x: 404 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon5.png'), y: 452 / scene.hauteur * 100, x: 457 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon6.png'), y: 430 / scene.hauteur * 100, x: 301 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon7.png'), y: 435 / scene.hauteur * 100, x: 450 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon8.png'), y: 434 / scene.hauteur * 100, x: 688 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon9.png'), y: 471 / scene.hauteur * 100, x: 605 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon10.png'), y: 450 / scene.hauteur * 100, x: 638 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon11.png'), y: 474 / scene.hauteur * 100, x: 402 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon12.png'), y: 448 / scene.hauteur * 100, x: 501 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon1.png'), y: 477 / scene.hauteur * 100, x: 316 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon2.png'), y: 443 / scene.hauteur * 100, x: 706 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon3.png'), y: 458 / scene.hauteur * 100, x: 308 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon4.png'), y: 477 / scene.hauteur * 100, x: 664 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon5.png'), y: 473 / scene.hauteur * 100, x: 654 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon6.png'), y: 445 / scene.hauteur * 100, x: 387 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon7.png'), y: 471 / scene.hauteur * 100, x: 440 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon8.png'), y: 476 / scene.hauteur * 100, x: 713 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon9.png'), y: 445 / scene.hauteur * 100, x: 231 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon10.png'), y: 455 / scene.hauteur * 100, x: 140 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon11.png'), y: 466 / scene.hauteur * 100, x: 750 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon12.png'), y: 456 / scene.hauteur * 100, x: 671 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon1.png'), y: 477 / scene.hauteur * 100, x: 795 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon2.png'), y: 464 / scene.hauteur * 100, x: 453 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon3.png'), y: 430 / scene.hauteur * 100, x: 855 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon4.png'), y: 463 / scene.hauteur * 100, x: 230 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon5.png'), y: 469 / scene.hauteur * 100, x: 328 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon6.png'), y: 475 / scene.hauteur * 100, x: 374 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon7.png'), y: 437 / scene.hauteur * 100, x: 474 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon8.png'), y: 476 / scene.hauteur * 100, x: 106 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon9.png'), y: 453 / scene.hauteur * 100, x: 619 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon10.png'), y: 462 / scene.hauteur * 100, x: 303 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon11.png'), y: 443 / scene.hauteur * 100, x: 243 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon12.png'), y: 448 / scene.hauteur * 100, x: 174 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon1.png'), y: 471 / scene.hauteur * 100, x: 632 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon2.png'), y: 439 / scene.hauteur * 100, x: 289 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon3.png'), y: 456 / scene.hauteur * 100, x: 346 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon4.png'), y: 474 / scene.hauteur * 100, x: 808 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon5.png'), y: 450 / scene.hauteur * 100, x: 281 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon6.png'), y: 433 / scene.hauteur * 100, x: 335 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon7.png'), y: 470 / scene.hauteur * 100, x: 751 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon8.png'), y: 443 / scene.hauteur * 100, x: 403 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon9.png'), y: 451 / scene.hauteur * 100, x: 832 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon10.png'), y: 450 / scene.hauteur * 100, x: 255 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon11.png'), y: 451 / scene.hauteur * 100, x: 299 / scene.largeur * 100 },
+  { image: require('tri/assets/bonbon12.png'), y: 450 / scene.hauteur * 100, x: 51 / scene.largeur * 100 }
 ];
 
 export default { pieces };
