+++ src/situations/tri/modeles/situation.js
@@ -1,8 +1,9 @@
 import SituationCommune from 'commun/modeles/situation';
+import Piece from './piece';
 
 export default class Situation extends SituationCommune {
   constructor ({ pieces }) {
     super();
-    this.pieces = pieces;
+    this.pieces = pieces.map((piece) => new Piece(piece));
   }
 }
