+++ src/situations/tri/vues/piece.js
@@ -1,12 +1,3 @@
-export default class VuePiece {
-  constructor (piece) {
-    this.piece = piece;
-  }
+import VuePieceCommune from 'commun/vues/piece';
 
-  affiche (pointInsertion, $) {
-    const $piece = $(`<img class="piece" src="${this.piece.image}">`);
-    $piece.css('top', this.piece.posY);
-    $piece.css('left', this.piece.posX);
-    $(pointInsertion).append($piece);
-  }
-}
+export default VuePieceCommune;
